# Made stashie ready for https://github.com/TehCheat/PoEHUD/pull/224/commits/d0bf3311eb093c33c5a601d9651534016c83c94b#diff-4b08d90b638a76d0aaac228a671000cd, it won't work until tehCheat has accepted that PullRequest, but when he has Stashie is gonna be so much better.
## Following changes are made:
Stashie now get's the index of the current visible stash from memory (thanks to Qvin for implementing IndexOfVisible stash into PoeHud)

## If the user has a dropdown menu (more than 30 tabs)
The method of switching tabs is now based on the travel distance (amount of left/right arrow presses) between the current visible tab and the tab we wish to go to.
If the distance is greater than 3, it will switch via the dropdown menu, otherwise it will switch via arrowkeys.